Move Window DebugActions to own file
(cherry picked from commit 76d62ca77d612c58393df0c516dbdf42e5a50d0e)
